Move interfaces to correct module
